fballoc and statum
update fballoc & statum to OpenPlatform spec
