+++ lib/fballoc/.vscode/fballoc.code-workspace
@@ -1,0 +1,18 @@
+{
+	"folders": [
+		{
+			"path": ".."
+		},
+		{
+			"path": "../../open"
+		},
+		{
+			"path": "../../bin"
+		},
+		{
+			"path": "../../lib"
+		}
+	],
+	"settings": {
+	},
+}
