+++ lib/statum/.vscode/statum.code-workspace
@@ -1,0 +1,18 @@
+{
+	"folders": [
+		{
+			"path": ".."
+		},
+		{
+			"path": "../../open"
+		},
+		{
+			"path": "../../bin"
+		},
+		{
+			"path": "../../lib"
+		}
+	],
+	"settings": {
+	},
+}
